# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Yudetter SNS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /explore/popular:
    get:
      operationId: getPopular
      tags: [explore]
      summary: Get popular yudates sorted by engagement
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Popular yudates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YudatePage"
  /explore:
    get:
      operationId: getExplore
      tags: [explore]
      summary: Get trending and recent yudates for explore page
      parameters:
        - name: cursor
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Explore yudates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YudatePage"
  /explore/search:
    get:
      operationId: search
      tags: [explore]
      summary: Search yudates and users
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [yudates, users, all, latest, popular, oldest]
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResults"

  # ── NOTIFICATIONS ─────────────────────────────────────────────────────────
...
  /games:
    get:
      operationId: getGames
      tags: [games]
      summary: Get list of YGS games
      responses:
        "200":
          description: List of games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Game"
    post:
      operationId: createGame
      tags: [games]
      summary: Create and publish a new game
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                htmlContent:
                  type: string
                playPrice:
                  type: integer
              required:
                - title
                - htmlContent
      responses:
        "201":
          description: Created game
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
  /games/{id}:
    get:
      operationId: getGame
      tags: [games]
      summary: Get a single game by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Game detail including HTML content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
    put:
      operationId: updateGame
      tags: [games]
      summary: Update a game by ID
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                htmlContent:
                  type: string
                playPrice:
                  type: integer
              required:
                - title
                - htmlContent
      responses:
        "200":
          description: Game updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
  /games/{id}/charge:
    post:
      operationId: chargeGamePayment
      tags: [games]
      summary: Process in-game YD payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: integer
                description:
                  type: string
              required:
                - amount
                - description
      responses:
        "200":
          description: Payment transaction ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  transactionId:
                    type: integer
                required:
                  - success

  # ── RANKINGS ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        setupComplete:
          type: boolean
        isPrivate:
          type: boolean
        isBlocking:
          type: boolean
        isBlockedBy:
          type: boolean
        pinnedYudateId:
          type: ["integer", "null"]
        headerUrl:
          type: ["string", "null"]
        birthday:
          type: ["string", "null"]
        id:
          type: integer
        clerkId:
          type: string
        username:
          type: string
        displayName:
          type: string
        email:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        followerCount:
          type: integer
        followingCount:
          type: integer
        yudateCount:
          type: integer
        isFollowing:
          type: boolean
        isFollowPending:
          type: boolean
        createdAt:
          type: string
          format: date-time
        visibility:
          type: string
          enum: [public, followers]
        isSpoiler:
          type: boolean
        yudedollar:
          type: integer
        badgeType:
          type: ["string", "null"]
        consecutiveLoginDays:
          type: integer
        rankingOptIn:
          type: boolean
      required:
        - id
        - clerkId
        - username
        - displayName
        - email
        - followerCount
        - followingCount
        - yudateCount
        - isFollowing
        - isFollowPending
        - isPrivate
        - isBlocking
        - isBlockedBy
        - setupComplete
        - createdAt
        - yudedollar
        - consecutiveLoginDays
        - rankingOptIn
...
    Yudate:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        author:
          $ref: "#/components/schemas/UserProfile"
        likeCount:
          type: integer
        reyudateCount:
          type: integer
        replyCount:
          type: integer
        isLiked:
          type: boolean
        isReyudated:
          type: boolean
        quotedYudate:
          oneOf:
            - $ref: "#/components/schemas/QuotedYudate"
            - type: "null"
        replyToId:
          type: ["integer", "null"]
        imageUrl:
          type: ["string", "null"]
        visibility:
          type: string
          enum: [public, followers]
        isSpoiler:
          type: boolean
        reactions:
          type: array
          items:
            $ref: "#/components/schemas/YudateReaction"
        superYudateAmount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - content
        - author
        - likeCount
        - reyudateCount
        - replyCount
        - isLiked
        - isReyudated
        - superYudateAmount
        - createdAt
    QuotedYudate:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        author:
          $ref: "#/components/schemas/UserProfile"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - content
        - author
        - createdAt
...
    YudatePage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Yudate"
        nextCursor:
          type: ["integer", "null"]
      required:
        - items

    # ── NOTIFICATION ─────────────────────────────────────────────────────────
...
    SearchResults:
      type: object
      properties:
        yudates:
          type: array
          items:
            $ref: "#/components/schemas/Yudate"
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
      required:
        - yudates
        - users
...
    YudateReaction:
      type: object
      properties:
        emoji:
          type: string
        count:
          type: integer
        isReacted:
          type: boolean
      required:
        - emoji
        - count
        - isReacted
...
    Game:
      type: object
      properties:
        id:
          type: integer
        creator:
          $ref: "#/components/schemas/UserProfile"
        title:
          type: string
        description:
          type: ["string", "null"]
        htmlContent:
          type: ["string", "null"]
        playPrice:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - creator
        - title
        - playPrice
        - createdAt
        - updatedAt
...